add additional check for sanity

+++ kantek/plugins/private/purge.py
@@ -1,4 +1,5 @@
 import logging
+from typing import Dict
 
 from telethon.tl.custom import Message
 from telethon.tl.types import Channel
@@ -10,7 +11,7 @@ tlog = logging.getLogger('kantek-channel-log')
 
 
 @k.command('purge', admins=True)
-async def puge(client: Client, chat: Channel, msg: Message, args, event) -> None:
+async def purge(client: Client, chat: Channel, msg: Message, args, kwargs: Dict, event) -> None:
     """Purge all messages from the the point the command was sent to the message that was replied to.
 
     Arguments:
@@ -19,6 +20,7 @@ async def puge(client: Client, chat: Channel, msg: Message, args, event) -> None
     Examples:
         {cmd}
     """
+    sure = kwargs.get('YesImSure', False)
     await msg.delete()
     if not chat.megagroup:
         return
@@ -36,4 +38,8 @@ async def puge(client: Client, chat: Channel, msg: Message, args, event) -> None
             message_ids = await client.get_messages(chat, min_id=reply_msg.id, max_id=msg.id)
         else:
             message_ids = list(range(reply_msg.id, msg.id))
-    await client.delete_messages(chat, message_ids)
+
+    if sure:
+        await client.delete_messages(chat, message_ids)
+    else:
+        await client.respond(purge, 'Please use -YesImSure to clarify you understand the extend of your actions!!')
